+++ resources/views/tasks/show.blade.php
@@ -1,31 +1,42 @@
 @extends('layouts.app')
 
 @section('content')
+    @if (Auth::check())
 
-<!-- ここにページ毎のコンテンツを書く -->
-    <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
+        <!-- ここにページ毎のコンテンツを書く -->
+        <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
     
-    <table class="table table-bordered">
-        <tr>
-            <th>id</th>
-            <td>{{ $task->id }}</td>
-        </tr>
-        <tr>
-            <th>ステータス</th>
-            <td>{{ $task->status }}</td>
-        </tr>
-        <tr>
-            <th>タスク</th>
-            <td>{{ $task->content }}</td>
-        </tr>
-    </table>
+        <table class="table table-bordered">
+            <tr>
+                <th>id</th>
+                <td>{{ $task->id }}</td>
+            </tr>
+            <tr>
+                <th>ステータス</th>
+                <td>{{ $task->status }}</td>
+            </tr>
+            <tr>
+                <th>タスク</th>
+                <td>{{ $task->content }}</td>
+            </tr>
+        </table>
     
-    {{-- タスク編集ページへのリンク --}}
-    {!! link_to_route("tasks.edit", "このタスクを編集", ["task" => $task->id], ["class" => "btn btn-light"]) !!}
+        {{-- タスク編集ページへのリンク --}}
+        {!! link_to_route("tasks.edit", "このタスクを編集", ["task" => $task->id], ["class" => "btn btn-light"]) !!}
     
-    {{-- タスク削除フォーム --}}
-    {!! Form::model($task, ["route" => ["tasks.destroy", $task->id ], "method" => "delete"]) !!}
-        {!! Form::submit("削除", ["class" => "btn btn-danger"]) !!}
-    {!! Form::close() !!}
+        {{-- タスク削除フォーム --}}
+        {!! Form::model($task, ["route" => ["tasks.destroy", $task->id ], "method" => "delete"]) !!}
+            {!! Form::submit("削除", ["class" => "btn btn-danger"]) !!}
+        {!! Form::close() !!}
+        
+    @else
+        <div class="center jumbotron">
+            <div class="text-center">
+                <h1>Welcome to the Microposts</h1>
+                {{-- ユーザ登録ページへのリンク --}}
+                {!! link_to_route('signup.get', 'Sign up now!', [], ['class' => 'btn btn-lg btn-primary']) !!}
+            </div>
+        </div>
+    @endif
 
 @endsection
